Move all parse`s classes to package parser
