+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/general/parser/parameterized_type/ListParameterizedType.java
@@ -1,0 +1,29 @@
+package ua.edu.chdtu.deanoffice.api.general.parser.parameterized_type;
+
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+import java.util.List;
+
+public class ListParameterizedType implements ParameterizedType {
+
+    private Type type;
+
+    public ListParameterizedType(Type type) {
+        this.type = type;
+    }
+
+    @Override
+    public Type[] getActualTypeArguments() {
+        return new Type[] {type};
+    }
+
+    @Override
+    public Type getRawType() {
+        return List.class;
+    }
+
+    @Override
+    public Type getOwnerType() {
+        return null;
+    }
+}
